Make the Settings side of the date format settings more locale-aware.
Remove the two formats that have a spelled-out month, since applications
using this setting are trying to format numeric dates.

Do not forcibly set the setting the first time you go into Date & Time --
let the setting remain null if it was null before.

Add a choice corresponding to null to the list of format options.  It will
look like "Normal (12-31-2009)" in the list, and will cause the system
to use whatever numeric format the locale calls for.

For the other choices, feed them to the locale-aware formatter so they
will appear with the punctuation that the locale calls for.