Redo power widget layout with the proper widget style.
The power widget now uses the fuzzy-edged widget background.
It has a resonable layout in portrait and landscape.
appwidget_inner is the background for a button on the inside, with a border on the left.
appwidget_outer is the background for a button on the outside (i.e. leftmost); it
does not have a border.