1. Added the missing timezone for mainland China "Asia/Shanghai". In each localized version of timezones.xml, the name of this timezone is defined like "China Standard Time (Beijing)" if some other timezones are also in this format, or just "Beijing" if all of other timezones are of city names.
2. Changed time zone name of "Asia/Taipei" from "China Standard Time (Taipei)"
to "Taipei" in some localized versions of timezones.xml.
"China Standard Time" is for China mainland, and the timezone of Taipei
is normally called "Taipei Standard Time" (or "Chungyuan Standard Time").
Though for now the offset of "Taipei Standard Time" and "China Standard Time"
is the same, they were different during 1986-1991 because of daylight saving
of China mainland.