Fix 2504863: Use password min/max constants in ChooseLockPassword
This fixes a bug where the system was defaulting to 8 digits max password length.

The UI is best for N <=12, but works for more.  I set the default limit to 16 which is
more than most humans can remember. That also seems to be the practical limit for the small
sample of DevicePolicyManagers I checked.

DevicePolicyManagers can always specify more digits if they really want.

Change-Id: I8637e0c9366800886759424b5a28f48dca15c535