Restore dependency between volume settings and silent mode.
Volume (of the ringer, notifications, media, and alarms) may
no longer be adjusted while the phone is in silent mode.
This was the behavior in Eclair and earlier releases.

This rolls back a small portion of change I724c43aa in the
interest of stability (see bugs listed below for weirdness
that occurs when the ringer's SeekBarVolumizer is visible
when the ringtone stream is muted).

Bug: 2552268
Bug: 2544706
Change-Id: I4858e6d85e5ec8c4bba83b649e24dd5915252eb2