Improve the add-device-admin layout.
Make it so that the app's explanation comes before the list of possible
policies. This gives the app a chance to explain itself instead of having
the explanation hidden below the fold. To avoid situations in which one
bunch of text items can obscure the other, use an expandable/ellipsizable
TextView with an indicator image. This resolves b/2992594.

Change-Id: I0eb0d0c46bb4be2ec2e019d741915537e1fcc592