Fixes in UsageBarPreference
The widget layout that was used resulted in a clipped usage bar.
It also had a 6dip left offset inherited from the default layout
which adds margin after the text (even if the text is empty here).
Used a custom layout for the whole preference to fix this issue.

Other fixes in the drawing code to prevent accumulative rounding
issues.

Change-Id: Ief9bf85260345355615c9670624d86f0893eb2d2