Adding an improve face matching settings option for Face Unlock
Adding an option which will launch a version of setup where faces
can be added to the current gallery.  It requires the user to first
confirm their password before launching addToSetup.

Patch 3 - Updated for renaming of FackLockTutorial to SetupIntro.
Now it is called every time regardless of it it's showing the tutorial
and a flag is passed in to determine whether or not SetupIntro shows
the tutorial.

Patch 4 - Removed 'Setup Complete' toast at the end of screen lock
setups since it was primarily for Face Unlock and the congratulations
screen removes the need for it.

Change-Id: Idc5f960809d992ec7bbce59ef1e13b95ef7cce45