WiFi: only send engine setting when key_id is set
For EAP methods that don't need a key, don't set the engine settings or
else wpa_supplicant will have an error. Instead, look to see whether
key_id is set and use that to indicate when the engine should be
set.

Additionally, don't require keystore when key_id is not null but is
still an empty string.

Bug: 6480500
Change-Id: I137de47dfcbbd3da24d693701e249e15d7528228