Fix 6654057: Fixing crash if device is rotated during confirm password
In onActivityResult, mBiometricWeakLiveliness can be null if the device was rotated
during confirm password.  To fix this, the line accessing mBiometricWeakLiveliness can
safely be removed.  It's unnecessary because when onActivityResult finishes, onResume is
always called when it displays the security settings again.  Since onResume sets the
checkmark value based on the stored value in lockPatternUtils, the check mark will always
show the correct value when settings is displayed.

Change-Id: Idba9ae1d829911a9581bb6678c4d1f3b44784c63