Fix incorrect updating of headers in Settings
Settings was overriding the header list in PreferenceActivity and
maintaining its own. In the case where Settings activity is restored
from saved state, the header lists went out of sync and couldn't
be updated anymore.

This makes use of a new private method in PreferenceActivity to
retrieve the header list object so that they stay in sync and can
be updated.

Also make sure to invalidate headers on resume, so that the latest
account types are loaded.

Bug: 6685701
Change-Id: I76ed1e1579a70fb5abfcea417d7e981907bdc02c