Fix Settings NPE
Labels for account types can be null which causes a NPE to occur in
Settings. If the label is null, don't bother adding an account
header. If we don't do this, then it would blow up when trying to
sort the headers anyway, so this change will fix the crash.

Bug 6727947

Change-Id: I284d7204b6d21c8d1b0a36e18ecae2c2361eeb84